+++ Editor/Inspector/SerializedLuaTableDrawer.cs
@@ -117,7 +117,7 @@ namespace LunyEditor.Inspector
 				var dupeKey = keyValuePair.Q<VisualElement>("duplicateKey");
 				if (keyField == null || keyText == null || dupeKey == null)
 					return;
-				
+
 				var isDuplicate = duplicateKeys.Contains(keyText.text);
 				dupeKey.style.display = isDuplicate ? DisplayStyle.Flex : DisplayStyle.None;
 			});