+++ CHANGELOG.md
@@ -6,6 +6,10 @@ https://github.com/CodeSmile-0000011110110111/de.codesmile.luny/issues
 See also the DevLog for a full history:
 https://docs.google.com/document/d/1jADnS6rsSx28-uWu8C4ubvCQD08zRO8sNLwtv8uASNU/edit?usp=sharing
 
+### 0.5.5
+
+- fixed: possible NullreferenceException in Script Context Table Inspector (harmless)
+
 ### 0.5.4
 
 - fixed: LunyRuntimeScript Inspector: 'Script Context' dictionary values could not be added